+++ app/(main)/custom/_components/change-color.module.css
@@ -9,6 +9,15 @@
     gap: 2rem;
 }
 
+.marimo_name {
+  width: 100%;
+
+  padding: 0.5rem;
+  font-size: 1.2rem;
+
+  text-align: center;
+}
+
 @media (max-width: 768px){
   .wrapper {
     flex-direction: column;
